+++ src/pages/studyView/StudyViewPageStore.ts
@@ -388,7 +388,7 @@ import { CustomDriverNumericGeneMolecularData } from 'shared/model/CustomDriverN
 import {
     PlotsColoringParam,
     PlotsSelectionParam,
-} from 'pages/resultsView/ResultsViewURLWrapper';
+} from 'shared/components/plots/PlotsTabUrlParameters';
 
 export const STUDY_VIEW_FILTER_AUTOSUBMIT = 'study_view_filter_autosubmit';
