+++ src/shared/components/plots/ScatterPlot.tsx
@@ -666,25 +666,25 @@ export default class ScatterPlot<
             this.props.data,
             d => this.props.highlightedSamples?.includes(d.sampleId)
         );
-        const highlightedDataBuckets = separateScatterData<D>(
-            highlightedData,
+        const sharedArgs = [
             ifNotDefined(this.props.fill, '0x000000'),
             ifNotDefined(this.props.stroke, '0x000000'),
             ifNotDefined(this.props.strokeWidth, 0),
             ifNotDefined(this.props.strokeOpacity, 1),
             ifNotDefined(this.props.fillOpacity, 1),
             ifNotDefined(this.props.symbol, 'circle'),
-            this.props.zIndexSortBy
+            this.props.zIndexSortBy,
+        ] as const;
+        // both data accept the same styling props
+        // the difference is that highlighted data are larger in size and have unique labels
+        // because of this, we separate each highlighted data point into its own bucket
+        const highlightedDataBuckets = separateScatterData<D>(
+            highlightedData,
+            ...sharedArgs
         );
         const unHighlightedDataBuckets = separateScatterDataByAppearance<D>(
             unHighlightedData,
-            ifNotDefined(this.props.fill, '0x000000'),
-            ifNotDefined(this.props.stroke, '0x000000'),
-            ifNotDefined(this.props.strokeWidth, 0),
-            ifNotDefined(this.props.strokeOpacity, 1),
-            ifNotDefined(this.props.fillOpacity, 1),
-            ifNotDefined(this.props.symbol, 'circle'),
-            this.props.zIndexSortBy
+            ...sharedArgs
         );
         // highlighted data points should appear in front of the other data points
         return [...unHighlightedDataBuckets, ...highlightedDataBuckets];
